Fix Potential Issue on OEMx Platform
[Issue Description]
Identified a potential issue on the OEMx platform.

[Resolution]
Fix Potential issues on OEMx platform

Package/Module: OneSiliconPkg

[Impacted Platform]
ALL

Hsd-es-id: 22021041597
Original commit date: Wed Mar 26 23:44:32 2025 -0700
Original commit hash: 4240742eba9a22259f63daad695801b7c11bad4e

+++ OneSiliconPkg/Fru/PtlPcd/Library/PtlPcdGpioNativeLib/PtlPcdGpioNativeLib.c
@@ -1844,7 +1844,7 @@ PtlPcdGpioConfigureIoStandbyState (
   VOID
   )
 {
-  UINT8                   SignalIndex;
+  UINT32                  SignalIndex;
   UINT32                  NativeSignalIndex;
   GPIOV2_PAD_MODE         GpioPadMode;
   EFI_STATUS              Status;
